Move live-play action chooser out of headless completion runner.
Keeps headlessMatchCompletionRunner focused on the completion loop while live AI chooser logic lives in livePlayActionChooser.js.
